refactor: DiModule 생성 및 Container 역할 변경
DiModule이 이전의 Container 역할을 하며,
Container는 Singleton 인스턴스들을 저장하는 저장소가 된다.
