+++ koala/src/main/java/com/bignerdranch/android/koala/Container.kt
@@ -1,3 +1,8 @@
 package com.bignerdranch.android.koala
 
-interface Container
+import kotlin.reflect.KClass
+
+object Container {
+
+    val instances = mutableMapOf<KClass<*>, Any?>()
+}
