+++ koala/src/main/java/com/bignerdranch/android/koala/DefaultDiModule.kt
@@ -1,0 +1,8 @@
+package com.bignerdranch.android.koala
+
+import android.content.Context
+
+class DefaultDiModule : DiModule {
+
+    override var context: Context? = null
+}
