+++ koala/src/main/java/com/bignerdranch/android/koala/DiModule.kt
@@ -1,0 +1,7 @@
+package com.bignerdranch.android.koala
+
+import android.content.Context
+
+interface DiModule {
+    var context: Context?
+}
